Review: apply suggestions
