+++ pkg/security/security_profile/secprofs.go
@@ -347,7 +347,7 @@ func (m *Manager) linkProfile(profile *profile.Profile, workload *tags.Workload)
 	profile.Instances = append(profile.Instances, workload)
 
 	// can we apply the profile or is it not ready yet ?
-	if profile.LoadedInKernel.Load() && slices.Contains(m.config.RuntimeSecurity.AnomalyDetectionEventTypes, model.SyscallsEventType) {
+	if profile.LoadedInKernel.Load() && m.isSyscallAnomalyEnabled {
 		m.linkProfileMap(profile, workload)
 	}
 }
@@ -378,7 +378,7 @@ func (m *Manager) unlinkProfile(profile *profile.Profile, workload *tags.Workloa
 	}
 
 	// remove link between the profile and the workload
-	if slices.Contains(m.config.RuntimeSecurity.AnomalyDetectionEventTypes, model.SyscallsEventType) {
+	if m.isSyscallAnomalyEnabled {
 		m.unlinkProfileMap(profile, workload)
 	}
 }
